Make the VM's sizes and positions hold where size_t or long is 32 bits (windows M1a)
A 32-bit VM would have hung growing the heap past 1 GiB of live data: the
doubling wrapped to 0. It now stops with "out of memory". The loader's bound
added a length read from the file to the position, which can wrap; it now
subtracts, and the loader reads the file to its end instead of asking ftell
for its size. bytes_allocated is 64 bits, so --count prints the same
everywhere, and --heap-size and --gc-stress refuse what is not a size that
fits, where they truncated it.

BinIO's positions went through ftell and fseek with a long, which is 32 bits
on both Windows targets. sys_ftell and sys_fseek join vm/sys.h: ftello and
fseeko on POSIX (with a 64-bit off_t on a 32-bit system too), _ftelli64 and
_fseeki64 on Windows, ftell and fseek in sys_none.c.

tests/vm/run-vm-tests.sh, part of make test, gives the VM bytecode files and
options written to overflow what is computed from them and wants a message
and status 2.

+++ vm/vm.h
@@ -120,7 +120,7 @@ typedef struct VM {
     size_t gc_count;
     int64_t gc_user_us;      /* processor time spent collecting, in microseconds */
     int64_t gc_sys_us;
-    size_t bytes_allocated;
+    uint64_t bytes_allocated;  /* not size_t: --count prints the same where it is 32 bits */
     uint64_t objects_allocated;
     uint64_t instructions;   /* executed so far */
     size_t gc_stress;        /* --gc-stress N: collect before every Nth allocation; 0 = off */
